api/parameters: bring Port enum back
Move Port from hub classes back to common parameters module.

This would have had too much impact on existing users, so undo it.

Notes if we are ever bringing it back again:
- The idea behind this was to ultimately allow for unique ports in a multi-hub scenario:

ev3 = EV3Brick()
ev3.Port.A # This is port A on this particular instance

This may still be added while leaving the existing Port enum in place, which would then always be the ports of the hub on which the code runs.