doc: add external link decoration to sidebar
External link decorations are already enabled in the documentation
content, but the RTD theme doesn't extend this to the sidebar. This
copies the css from the RTD theme and adds it to the sidebar.

This makes it less of a surprise when you click on a link and find
yourself on a completely different website.