WIP: possible solution for awaitable functions
In our proposed multitasking framework, many blocking functions will
become awaitable when running in an async run loop. So we need a way
to give code completion tools to work in either case.

This introduces a pattern of a `MaybeAsync<Type>` where `MaybeAsync<Type>`
inherits from both `<Type>` and `typing.Awaitable[<Type>]`. This way
we get mostly proper completions for both `x = instance.method()` and
`x = await instance.method()`.

As a bonus, it is also a nice way to very quickly know if a method
should be awaited or not by just looking at the signature and not
having to read the full documentation.

Note: I tried a generic:

```python
class MaybeAsync(_T, Awaitable[_T]): ...

class ...

    def method(self) -> MaybeAsync[int]: ...
```

But the code completion engines don't seem to be able to figure that
one out, so it seems we will have to make our own type for each
return type.